requirements.txt
python-telegram-bot==20.3
solders==0.24.0
solana==0.27.0
python-dotenv